requirements.txt
aws-cdk-lib==2.66.0
constructs>=10.0.0,<11.0.0
attr==0.3.2
attrs==22.2.0
boto3==1.26.109
requests==2.28.2
Jinja2==3.1.2
jsii==1.80.0
jwt_generator==0.12.1
PyYAML==6.0
PyJWT==2.6.0

cryptography~=40.0.1
MarkupSafe~=2.1.2
certifi~=2022.12.7
six~=1.16.0
idna~=3.4
urllib3~=1.26.15
pycparser~=2.21
botocore~=1.29.113